+++ src/pages/admin/components/docs/AdminConfirmTable.tsx
@@ -2,11 +2,18 @@ import { useState, useEffect } from 'react';
 
 import Pagination from '@gdsc/components/common/pagination/pagination';
 
-import { applyDocsInterface } from '@gdsc/apis/hooks/admin/useGetApplyDocs';
+import { applyDocsInterface } from '@gdsc/apis/hooks/admin/docs/useGetApplyDocs';
 import { useGetSearch } from '@gdsc/apis/hooks/admin/useGetSearch';
 
+import {
+  StyledTable,
+  TableHeader,
+  TableHeaderRow,
+  TableCell,
+  TableHeaderCell,
+  TableRow,
+} from '../MemberTable.style';
 import { columns } from './AdminTableDocs';
-import styled from '@emotion/styled';
 import { MemberData, Track } from '@gdsc/types/AdminInterface';
 import {
   flexRender,
@@ -22,63 +29,6 @@ type Props = {
   currentPage: number;
   setCurrentPage: (page: number) => void;
 };
-
-const StyledTable = styled.table`
-  width: 100%;
-  border-collapse: collapse;
-  margin: 30px 0px;
-`;
-
-const TableHeader = styled.thead`
-  height: 59px;
-  background-color: rgba(255, 255, 255, 0.7);
-  color: var(--color-halti);
-  white-space: nowrap;
-  text-align: left;
-`;
-
-const TableHeaderRow = styled.tr`
-  border-bottom: none;
-`;
-
-const TableCell = styled.td`
-  height: 59px;
-  border: none;
-  white-space: nowrap;
-  font-size: 14px;
-
-  &:first-of-type {
-    border-top-left-radius: 12px;
-    border-bottom-left-radius: 12px;
-    padding-left: 10px;
-  }
-
-  &:last-of-type {
-    border-top-right-radius: 12px;
-    border-bottom-right-radius: 12px;
-    padding-right: 10px;
-  }
-`;
-
-const TableHeaderCell = styled.th`
-  border: none;
-  width: 1rem;
-  &:first-of-type {
-    border-top-left-radius: 12px;
-    border-bottom-left-radius: 12px;
-  }
-
-  &:last-of-type {
-    border-top-right-radius: 12px;
-    border-bottom-right-radius: 12px;
-  }
-`;
-
-const TableRow = styled.tr`
-  &:nth-of-type(even) {
-    background-color: rgba(255, 255, 255, 0.15);
-  }
-`;
 
 const AdminConfirmTable = ({
   data,
